Fix by pr comments
